blob sidecar listener logic in a separate class
